+++ CHANGELOG.md
@@ -4,6 +4,18 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+### Added
+
+- Allow setting a device for the instance instead of per request
+- Allow passing Axios options for all requests
+- Allow passing a bearer token for all requests
+
+### Changed
+
+- Make the bearer token optional when creating the instance
+- Use the same configuration object for creating an instance and for all
+  requests
+
 ## [0.1.0] - (2021-05-04)
 
 ### Added